+++ src/devwaveformStream.c
@@ -172,10 +172,10 @@ static long writeData(dbCommon *record, format_t *format)
                         break;
 #ifdef DBR_INT64
                     case DBF_INT64:
-                        dval = ((epicsInt64 *)wf->bptr)[nowd];
+                        dval = (double)((epicsInt64 *)wf->bptr)[nowd];
                         break;
                     case DBF_UINT64:
-                        dval = ((epicsUInt64 *)wf->bptr)[nowd];
+                        dval = (double)((epicsUInt64 *)wf->bptr)[nowd];
                         break;
 #endif
                     case DBF_LONG:
@@ -215,10 +215,10 @@ static long writeData(dbCommon *record, format_t *format)
                 {
 #ifdef DBR_INT64
                     case DBF_INT64:
-                        lval = ((epicsInt64 *)wf->bptr)[nowd];
+                        lval = (long)((epicsInt64 *)wf->bptr)[nowd];
                         break;
                     case DBF_UINT64:
-                        lval = ((epicsUInt64 *)wf->bptr)[nowd];
+                        lval = (long)((epicsUInt64 *)wf->bptr)[nowd];
                         break;
 #endif
                     case DBF_LONG:
